Adiciona arquivo componente InputOption
